creating reusable snackbar component
