+++ moped-editor/src/components/DataGridPro/DataGridErrorSnackbar.js
@@ -1,0 +1,27 @@
+import React from "react";
+
+import { Snackbar, Alert } from "@mui/material";
+
+const DataGridErrorSnackbar = ({ snackbarState, setSnackbarState }) => {
+  /**
+   * Return Snackbar state to default, closed state
+   */
+  const handleSnackbarClose = () => {
+    setSnackbarState(DEFAULT_SNACKBAR_STATE);
+  };
+
+  return (
+    <Snackbar
+      anchorOrigin={{ vertical: "top", horizontal: "right" }}
+      open={snackbarState.open}
+      onClose={handleSnackbarClose}
+      key={"datatable-snackbar"}
+    >
+      <Alert onClose={handleSnackbarClose} severity={snackbarState.severity}>
+        {snackbarState.message}
+      </Alert>
+    </Snackbar>
+  );
+};
+
+export default DataGridErrorSnackbar;
